Proceed with files decoupling
