+++ src/tinymce/callbacks/index.js
@@ -1,0 +1,2 @@
+export { file_picker_callback } from "./file_picker_callback.js";
+export { init_instance_callback } from "./init_instance_callback.js";
